pyproject.toml
[tool.poetry]
name = "creatifyinterview"
version = "0.1.0"
description = ""
authors = [
    {name = "wusp"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.2.0"
djangorestframework = "^3.15.0"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.0"
mysqlclient = "^2.2.0"
python-decouple = "^3.8"
aiohttp = "^3.9.0"
asgiref = "^3.7.0"
uvloop = "^0.19.0"
gunicorn = "^21.2.0"
uvicorn = "^0.24.0"

[tool.poetry.group.dev.dependencies]
# 这里可以添加开发依赖

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"